added global Loader, use loader redux
1) change component Loading to Loader
2) delete local loader from WatercardList component
3) delete local loader from Calendar component
4) add global Loader to main.js
5) delete loader from App.jsx
