+++ src/App.jsx
@@ -1,12 +1,10 @@
 import { lazy, useEffect } from "react";
+import { useDispatch } from "react-redux";
 import { Route, Routes } from "react-router-dom";
 import { SharedLayout } from "./components/SharedLayout/SharedLayout";
-import { useDispatch, useSelector } from "react-redux";
-import { selectIsRefreshing } from "./redux/user/selectors";
 import { fetchCurrentUser } from "./redux/user/userOps";
 import { RestrictedRoute } from "./components/RestrictedRoute";
 import { PrivateRoute } from "./components/PrivateRoute";
-import { Loading } from "./components/Loading/Loading";
 
 const HomePage = lazy(() => import("./pages/HomePage/HomePage"));
 const SignInPage = lazy(() => import("./pages/SignInPage/SignInPage"));
@@ -23,17 +21,11 @@ const ResetPasswordPage = lazy(() =>
 export function App() {
   const dispatch = useDispatch();
 
-  const isRefreshing = useSelector(selectIsRefreshing);
-
   useEffect(() => {
     dispatch(fetchCurrentUser());
   }, [dispatch]);
 
-  return isRefreshing ? (
-    <div className="loader">
-      <Loading />
-    </div>
-  ) : (
+  return (
     <div>
       <Routes>
         <Route path="/" element={<SharedLayout />}>
@@ -64,7 +56,7 @@ export function App() {
                 component={<SignInPage />}
               />
             }
-            />
+          />
           <Route
             path="/tracker"
             element={
