+++ src/components/Calendar/Calendar.jsx
@@ -4,9 +4,7 @@ import CalendarItem from "../CalendarItem/CalendarItem";
 import {
   selectMonthWater,
   selectCalendarMonth,
-  selectWaterLoading,
 } from "../../redux/water/selectors";
-import { Loading } from "../Loading/Loading";
 import { calculateDaysInMonth } from "../../helpers/calculateDaysInMonth ";
 import styles from "./Calendar.module.css";
 import { selectDesiredVolume } from "../../redux/user/selectors";
@@ -15,21 +13,12 @@ const Calendar = () => {
   const [activeDay, setActiveDay] = useState(null);
   const monthWater = useSelector(selectMonthWater);
   const currentMonth = useSelector(selectCalendarMonth);
-  const isLoading = useSelector(selectWaterLoading);
-
   const desiredVolume = useSelector(selectDesiredVolume);
 
   const daysInMonth = useMemo(
     () => calculateDaysInMonth(currentMonth),
     [currentMonth]
   );
-
-  if (isLoading)
-    return (
-      <div className={styles.loading}>
-        <Loading customHeight="124" />
-      </div>
-    );
 
   return (
     <ul className={styles.calendarList}>
@@ -52,7 +41,7 @@ const Calendar = () => {
         const isActive = formattedDate === activeDay;
 
         const handleSelect = () => {
-          setActiveDay(formattedDate); // Устанавливаем активный день
+          setActiveDay(formattedDate);
         };
 
         return (
