+++ src/main.jsx
@@ -8,6 +8,7 @@ import { persistor, store } from "./redux/store.js";
 import { PersistGate } from "redux-persist/integration/react";
 import "./index.css";
 import { GoogleOAuthProvider } from "@react-oauth/google";
+import Loader from "./components/Loader/Loader.jsx";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
@@ -16,6 +17,7 @@ ReactDOM.createRoot(document.getElementById("root")).render(
         <BrowserRouter>
           <HelmetProvider>
             <GoogleOAuthProvider clientId="416780880266-5n7irv3l8fmjbbi7v56vafihee55k2eo.apps.googleusercontent.com">
+              <Loader />
               <App />
             </GoogleOAuthProvider>
           </HelmetProvider>
